Refactoring logging to better support MCP server
