+++ src/utils.ts
@@ -12,10 +12,7 @@ import * as open from "open";
 import * as ora from "ora";
 import * as process from "process";
 import { Readable } from "stream";
-import * as winston from "winston";
-import { SPLAT } from "triple-beam";
 import { AssertionError } from "assert";
-import { stripVTControlCharacters } from "node:util";
 import { getPortPromise as getPort } from "portfinder";
 
 import { configstore } from "./configstore";
@@ -469,22 +466,6 @@ export async function promiseProps(obj: any): Promise<any> {
 }
 
 /**
- * Attempts to call JSON.stringify on an object, if it throws return the original value
- * @param value
- */
-export function tryStringify(value: any) {
-  if (typeof value === "string") {
-    return value;
-  }
-
-  try {
-    return JSON.stringify(value);
-  } catch {
-    return value;
-  }
-}
-
-/**
  * Attempts to call JSON.parse on an object, if it throws return the original value
  * @param value
  */
@@ -497,34 +478,6 @@ export function tryParse(value: any) {
     return JSON.parse(value);
   } catch {
     return value;
-  }
-}
-
-/**
- *
- */
-export function setupLoggers() {
-  if (process.env.DEBUG) {
-    logger.add(
-      new winston.transports.Console({
-        level: "debug",
-        format: winston.format.printf((info) => {
-          const segments = [info.message, ...(info[SPLAT] || [])].map(tryStringify);
-          return `${stripVTControlCharacters(segments.join(" "))}`;
-        }),
-      }),
-    );
-  } else if (process.env.IS_FIREBASE_CLI) {
-    logger.add(
-      new winston.transports.Console({
-        level: "info",
-        format: winston.format.printf((info) =>
-          [info.message, ...(info[SPLAT] || [])]
-            .filter((chunk) => typeof chunk === "string")
-            .join(" "),
-        ),
-      }),
-    );
   }
 }
